# docs/openapi/tessera.yaml
...
openapi: 3.1.0
info:
  title: Tessera viewer and session planes
  version: "1"
  summary: A permission-masked point service — the two planes a client speaks to.
  description: |
    Two listeners. The **session plane** mints and revokes viewer sessions and is authenticated by
    the operator-configured *session credential*, a shared secret; it is never browser-facing. The
    **viewer plane** answers about the corpus as one principal may see it and is authenticated by
    the *session token* the session plane minted. `/healthz` and `/readyz` sit on both, unauthenticated,
    bare status codes with no body.

    `api_version` is published by `GET /v1/meta` and by nothing else. There is no version header,
    on either side, so a client must not send one as a negotiation and must not expect a refusal
    if its assumed version is wrong (contracts §3.1).

    **Every error is one envelope** (`Error` below) with a closed code list, and the codes group
    into three refusal classes that tell a client whether to retry (contracts §3.1, design §10.6):

    - **Failure** — `500 fail-closed`, `503 not-ready`. Fail-closed and never a partial result.
    - **Backpressure** — `429 backpressure`. Retry the request *unchanged*, after `Retry-After`,
      which every 429 carries together with an agreeing body `retry_after_s`. The number is
      per-subject — `1` for the viewer and session planes' compute-admission gate, derived
...

...

...
servers:
  - url: http://{viewer-host}
...
    variables:
...
paths:
...
  /v1/viewport:
    post:
      tags: [viewer]
      operationId: viewport
      summary: Masked counts per tile, a served sample of points, and the served annotation artifacts, over a region at a depth.
      description: |
        The response is **not JSON**. It is a sequence of tagged, length-prefixed frames —
        `u8 kind`, `u32 little-endian payload length`, payload — each payload a complete Arrow IPC
        stream (JSON for the trailer), decodable alone (contracts §5). Kinds: `1` tiles (exactly
        one, first), `2` sub-cells (exactly one iff `underlay_offset` was supplied and non-zero),
        `5` artifacts (at most one, after tiles and before any points; **absent, never empty, when
        nothing is served**, deliberately), `3` points (zero or more, whole tiles per frame,
        boundaries not contract), `4` trailer (exactly one, last — its presence is the completeness
        signal). A body without a trailer is incomplete whatever the transport said; an unknown
        kind is a decoder error, never skipped. `docs/openapi/README.md` walks a decode in Python
        and JavaScript, and contracts §3.2 gives every batch's schema.

        **Exactly one of `bbox` and `tiles`.** `k` defaults to the deployment's `k_max_marks` and
        is capped by `max_k` (both in `/v1/meta`'s `selection`); it is design §7.2's cap clause,
        so a tile may return fewer than `k` while more are visible — that is the density signal,
        not truncation. **`k = 0` is the counts-only request**: tiles and artifacts as at any `k`,
        and no points frame at all. **A client must not decrease `k` as it zooms in**; the server
        cannot enforce this.

        The *artifacts* frame's `matched` column answers the request's `filters` per
        artifact, and is null where the request carried none — see `filters`. Its `layer` column
        is **dictionary-encoded** (u16 keys over utf8 values); its `rung` column is the number a
        client draws by — the declared level on a levelled layer, the response-local parent-chain
        depth on a treed one (the depth in the forest this response's own `parent_id` links form,
        after the budget cut), `0` on a flat one; and its two shape columns `shape_x`/`shape_y` —
        `list<list<list<uint32>>>`, parts then rings then vertices, one drawn geometry per artifact
        of the kind its layer's `shape` names — **trail the fixed columns and are absent from the
        schema entirely when no served layer draws a shape or the request did not ask for one** —
        an absent column is distinguishable from a null one, so a per-row null keeps its one
        meaning: *this layer declares no such property*, never *withheld* (decision 0076).
        `artifact_rows: "identity"` answers with the **same rows** in a fixed five-column schema —
        see its property.

        A request carrying a `highlight` gains three columns and loses none: `highlighted` fifth on
        the tiles frame (`uint64`, **always present**, equal to `matched` where the request carried
        no highlight), `highlighted` on the points frame (`bool`, after the render scalars and
        before the membership columns, **present only with a highlight**), and `highlighted`
        fifteenth on the artifacts frame (`bool?`, after `matched`). The served set is identical
        with and without one — see `highlight`.

        The artifacts frame's sixteenth column is `target` (`uint64?`): the `tessera_id` of the
        artifact this row is attached to, which for a clustering's label is its cluster. The
        target is a row of the same frame, and the column is null for a row attached to nothing.
        A label's `masked_count` is its own. The `"identity"` projection omits `target`.

        `layers` decides whether an artifacts frame can appear — see its schema; the semantics
        there are the owner's ruling of 2026-08-25 and the server change is landing on the s3
        track, so until that merges an *omitted* `layers` is read as `"all"` and the string form is
        not accepted. `artifact_budget` is honoured structurally (ancestors in place of their
        descendants), never by sampling, and is inert on a flat layer.

        Behind the compute-admission gate.
      security:
        - sessionToken: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ViewportRequest"
      responses:
        "200":
          description: The framed body. Headers carry the coordinates a client caches and revalidates by.
          headers:
            etag:
              description: The **content key**, quoted — identity, the served watermark, the overlay version, a per-process nonce (`delta-serving.md` §2). What a held band may be *declared* under. Echoed as `If-Match`; a mismatch yields the full response, never `412`.
              required: true
              schema:
                type: string
            x-tessera-identity-key:
              description: The **identity key** — idset, auth-data hash, mask-fragment identity, view. What a held band may be *rendered* under at all; the client's cache partition key.
              required: true
              schema:
                type: string
            x-tessera-pin:
              description: The generation stamp of the geometry this response was computed from, as JSON `{"prefix", "segments_version"}`. A stamp, not a selector — echo it back as `pin`, and the only effect is `x-tessera-stale`.
              required: true
              schema:
                type: string
            x-tessera-stale:
              description: "`1` if geometry has moved since the `pin` this request presented, else `0`. The **broadcast** form: it moves at every flush, merge and compaction whether or not anything this principal can see changed. Advisory in both directions."
              required: true
              schema:
                type: string
                enum: ["0", "1"]
            x-tessera-region:
              description: "Present exactly when the request's `filters` carried a `region` leaf: `exact` when every region's answer is exact for the shape against each point's stored position, or `cover; depth=<d>` when a shape's perimeter exceeded `max_region_cells` and the answer is exact for a cover of it — a superset — taken at depth `d` (`selection-operand.md` §6). A function of the shape and the grid alone, never of the rows, so it says nothing about the corpus; with several region leaves, the coarsest. A leaf by artifact is always exact."
              required: false
              schema:
                type: string
                pattern: "^(exact|cover; depth=[0-9]+)$"
            x-tessera-server-us:
              description: Server compute after admission to first-flush-ready, microseconds. The whole-stream wall total is the trailer's `stream_us`.
              required: true
              schema:
                type: string
                pattern: "^[0-9]+$"
            x-tessera-admission-us:
              description: Time spent in the compute-admission gate, microseconds.
              required: true
              schema:
                type: string
                pattern: "^[0-9]+$"
          content:
            application/octet-stream:
              schema:
                type: string
                format: binary
                description: Framed Arrow IPC streams and one JSON trailer, per contracts §5.
        "401":
          $ref: "#/components/responses/BadCredential"
        "403":
          $ref: "#/components/responses/ExpiredToken"
        "404":
          description: Unknown `view`.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "422":
          description: "`zoom` above 16; neither or both of `bbox`/`tiles`; a non-finite or inverted bbox; a tile prefix with bits above `zoom`; an `underlay_offset` beyond the deployment's ceiling, the grid or the sub-cell budget (refused, never clamped); an unknown filter column or an operator outside its family (an unknown *value* is an empty operand, never a refusal); `none_of` naming a text column or naming two columns."
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "429":
          $ref: "#/components/responses/Backpressure"
        "500":
          $ref: "#/components/responses/FailClosed"
        "503":
          $ref: "#/components/responses/NotReady"
  /v1/items:
    post:
      tags: [viewer]
      operationId: items
      summary: Every item this principal may see in one view that matches a filter, page by page, with the fields the caller names.
      description: |
        The response is **not JSON**. It is the viewport's framing: a sequence of frames, each
        `u8 kind`, `u32` little-endian payload length, payload. Kinds:

        - `6` head: JSON `ItemsHead`, exactly one, first.
        - `7` records: one Arrow IPC stream holding one batch of up to `page_rows` rows. Zero or
          more, each followed by a page end.
        - `8` page end: JSON `ItemsPageEnd`, carrying the cursor to resume after that page.
        - `4` trailer: JSON `ItemsTrailer`, exactly one, last.

        A body without a trailer is incomplete. A client whose connection drops resumes from the
        last page end it received and discards a records frame that no page end follows. A
        response with no rows is a head and a trailer. An unknown kind is a decoder error.
        `docs/openapi/README.md` walks a decode in Python and JavaScript.

        **Reading a whole result.** Each response carries at most `pages` pages and runs within
        the deployment's byte and time budgets. The trailer's `next` is the cursor to pass back
        as `cursor`, unchanged, until it is null. The server keeps nothing between requests: every
        page is built from the latest data, with the visible set composed again, so a deletion or
        suppression accepted during a read applies from the next page. A read returns each row at
        most once. A row inserted behind the cursor during a read is not returned; one inserted
        ahead of it is. `fields`, `system_fields`, `filters`, `keep_unmatched` and `page_rows` may
        change between the requests of one read, and the page schema changes with them.

        **Columns.** `tessera_id` (`uint64`) is first, then the named fields in the order named,
        then the system fields in the order named, then `tessera:matched` (`bool`) under
        `keep_unmatched`. Every named field is present whether or not an item carries a value. A
        number or bool arrives as its declared type, `timestamp_us` as `timestamp[us, UTC]`, a
        keyword or text field as `utf8`, and a category as `dictionary<int32, utf8>` holding its
        value keys, each frame's dictionary holding only the keys its rows carry. An absent value
        is null. `position` is two `float64` columns, `tessera:x` and `tessera:y`: the stored
        position converted back through the view's projection and frame, so degrees for a
        geographic view, exact to one step of the view's 32-bit grid. `external_id` is `binary`
        as `tessera:external_id`, null where the item has none. `labels` is `list<utf8>` as
        `tessera:labels`: the item's labels this principal holds, sorted. A rendered field is read
        from `view`; the item card reads the first view the principal can reach, so the two can
        differ. An item joined into `view` before its own ingest is flushed has null record
        fields until that flush.

        **Group-scoped fields** resolve as a filter leaf on them does: read in `view` under a view
        of the field's group, or of a group sharing its views, and otherwise pinned as
        `<field>@<key>`.

        **Order.** `map` returns rows by map cell in `view`, then by `tessera_id`. `stored`
        returns them in the order the record store holds items, which groups items by their full
        set of access labels. The two orders return the same rows. With neither `order` nor
        `cursor`, the server takes `stored` when any named field's `homes` in `/v1/meta` is
        `["record"]`, and `map` otherwise; that choice may change. A field read only from the
        record store is 1.3 to 1.8 times slower to read in map order at 100,000-row pages, and 8
        to 17 times slower at 1,000-row pages (measured).

        **Pages.** A page ends at `page_rows` rows, or before the row that would take its Arrow
        bytes past `selection.max_page_bytes`; a single larger row is sent alone. A page holds at
        least one row whenever one remains. When the response's time budget runs out, or the
        stream deadline cancels it, a page holding rows is sent short and the trailer follows.
        Every response moves the cursor on, including one that found no row under a sparse
        filter, and a response can run past its time budget by one evaluation of the filter over
        a stretch and one chunk of a page's scan.

        **Cursor.** Opaque, encrypted and authenticated under a key derived from the identity
        key, and bound to the route, the view and its incarnation, the order and the credential
        that received it. A cursor that fails any of these is one `422` with one detail. A
        cursor issued under another idset is the `409` below. A cursor survives flushes, merges,
        compactions and restarts, and stops opening when the identity key is rotated.

        On the bulk-read admission gate (`serve.bulk_admission`), whose queue is shed as `429`
        like the viewport's, so a long read never holds a slot the viewport and item routes use.
      security:
        - sessionToken: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ItemsRequest"
      responses:
        "200":
          description: The framed body. The headers are the viewport's, without its cache coordinates.
          headers:
            x-tessera-identity-key:
              description: The viewport's identity coordinate for this principal and view, equal to the one a viewport over the same view returns.
              required: true
              schema:
                type: string
            x-tessera-region:
              description: "Present exactly when the request's `filters` carried a `region` leaf, with the viewport's values: the coarsest verdict the head's count and the first page reached."
              required: false
              schema:
                type: string
                pattern: "^(exact|cover; depth=[0-9]+)$"
            x-tessera-server-us:
              description: Server time from admission to the head, microseconds. The whole response's is the trailer's `stream_us`.
              required: true
              schema:
                type: string
                pattern: "^[0-9]+$"
            x-tessera-admission-us:
              description: Time spent in the bulk-read admission gate, microseconds.
              required: true
              schema:
                type: string
                pattern: "^[0-9]+$"
          content:
            application/octet-stream:
              schema:
                type: string
                format: binary
                description: A head, records frames each followed by a page end, and a trailer, per contracts §5.
        "401":
          $ref: "#/components/responses/BadCredential"
        "403":
          $ref: "#/components/responses/ExpiredToken"
        "404":
          description: "Unknown `view`, or a pin naming no view of its field's group that this principal can reach."
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "409":
          description: "`idset` was supplied and differs from the deployment's, or the cursor was issued under another idset. Re-resolve by external id."
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "422":
          description: "The viewport's filter refusals; a body with an unknown key, a missing `view` or `fields`, or an `order` or `compression` outside its values; an unknown or repeated field; an unpinned group-scoped field outside its group, a pin on a field that is not group-scoped, or a group-scoped `text` field; an unknown system field; `page_rows` or `pages` of 0; `count` with a `cursor`; a cursor that does not open for this request, or whose order differs from `order`. `page_rows` above `selection.max_page_rows` is capped, not refused."
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "429":
          $ref: "#/components/responses/Backpressure"
        "500":
          $ref: "#/components/responses/FailClosed"
  /v1/items/{tessera_id}:
    post:
      tags: [viewer]
      operationId: item
      summary: One item's whole record, if this principal may see it.
      description: |
        The record assembled from its three homes and keyed by declared column name; a category
        appears as its vocabulary **key**, never its code; every width lands on a JSON number. **An
        absent field is absent from the object, never `null`.** `external_id` is the caller's own,
        base64, present only where one exists — the only place a caller external id appears on
        this plane.

        **The whole permitted picture of one point** (owner ruling 2026-09-01): beside the record,
        `views` names every view the item is in **that this principal may reach**, with the
        position each view places it at, and `scoped` carries the group-scoped attribute values
        keyed by the group's key. Both are gate-filtered against the session's visible-view set, so
        a point held only in views this principal cannot reach serves exactly what it served
        before — an empty `views` and no scoped entries — and nothing on this route names a
        gate-failed view.

        `404 unknown` is returned identically for *no such id* and *exists but not visible* — same
        status, same code, same detail, and the same work. Behind the compute-admission gate.
      security:
        - sessionToken: []
      parameters:
        - $ref: "#/components/parameters/TesseraId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ItemRequest"
      responses:
        "200":
          description: The record.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ItemResponse"
        "401":
          $ref: "#/components/responses/BadCredential"
        "403":
          $ref: "#/components/responses/ExpiredToken"
        "404":
          $ref: "#/components/responses/Unknown"
        "409":
          description: "`idset` was supplied and differs from the deployment's — *stale idset; re-resolve by external_id*. Decided before inversion and identically for every identifier."
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "429":
          $ref: "#/components/responses/Backpressure"
        "500":
          $ref: "#/components/responses/FailClosed"
...
components:
  securitySchemes:
    sessionCredential:
      type: http
      scheme: bearer
      description: The operator-configured session credential (`serve.session_credential`), a shared secret. Compared in constant time. Never handed to a browser.
    sessionToken:
      type: http
      scheme: bearer
      description: The `token` `POST /session/authorise` minted. A missing, malformed or unrecognised token — including a revoked one — is `401 bad-credential`; a recognised token past `expires_at` is `403 expired-token`. Both end the session.
  parameters:
    TesseraId:
      name: tessera_id
      in: path
      required: true
      description: The item's or artifact's opaque identifier, a `u64` in decimal. A keyed blinding permutation of the internal identity — stable across sessions, meaningless without the key, and **not** a cryptographic guarantee (contracts §5, decision 0014).
      schema:
        type: string
        pattern: "^[0-9]+$"
  responses:
    BadCredential:
      description: "`401 bad-credential`: missing, malformed or unrecognised bearer credential. On a token that previously worked this is the session ending — revoked — and a client treats it exactly as `403`."
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    ExpiredToken:
      description: "`403 expired-token`: the token is recognised and its `expires_at` has passed. Re-authorise."
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
...
    Backpressure:
      description: "`429 backpressure`: shed whole, never a narrowed answer. Retry unchanged after `Retry-After`, which agrees with the body's `retry_after_s`."
      headers:
        Retry-After:
          required: true
          description: Seconds. `1` for the compute-admission gate and the single-flight builders; other subjects derive their own figure. Read it rather than assuming it.
          schema:
            type: string
            pattern: "^[0-9]+$"
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/BackpressureError"
    FailClosed:
      description: "`500 fail-closed`: a mask, composition or containment failure. Partial results do not exist."
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    NotReady:
      description: "`503 not-ready`: unverified bundle, unready worker, unloaded plugin. No `Retry-After`, deliberately."
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
  schemas:
    Error:
      type: object
      description: The one error envelope, every plane (contracts §3.1). The code list is closed.
      additionalProperties: false
      required: [error, detail]
      properties:
        error:
          type: string
          enum:
            - bad-credential
            - expired-token
            - unknown
            - conflict
            - contract
            - backpressure
            - fail-closed
            - not-ready
        detail:
          type: string
          description: Human-readable. Identical for every branch of a `404`, so it discloses nothing an enumeration could use.
        retry_after_s:
          type: integer
          minimum: 0
          description: Present on every `429` and on nothing else; agrees with the `Retry-After` header.
...
    ItemRequest:
      type: object
      additionalProperties: false
      properties:
        pin:
          $ref: "#/components/schemas/Pin"
        idset:
          type: integer
          minimum: 0
          description: The idset the caller minted its `tessera_id` under. Supplying it makes a `tessera_id` from a past idset a `409` rather than another item's record.
...